Braking apart plots for more clear grouping of functions.
