+++ src/pyciemss/visuals/histogram.py
@@ -1,0 +1,83 @@
+from typing import List, Callable
+
+from numbers import Number
+
+import pandas as pd
+import numpy as np
+
+from . import vega
+
+
+def sturges_bin(data):
+    """Determine number of bin susing sturge's rule.
+    TODO: Consider Freedman-Diaconis (larger data sizes and spreads)
+    """
+    return int(np.ceil(np.log2(len(data))) + 1)
+
+
+def histogram_multi(
+    *,
+    xrefs: List[Number] = [],
+    yrefs: List[Number] = [],
+    bin_rule: Callable = sturges_bin,
+    return_bins: bool = False,
+    **data,
+) -> vega.VegaSchema:
+    """
+    Create a histogram with server-side binning.
+
+    TODO: Maybe compute overlap in bins explicitly and visualize as stacked?
+          Limits (practically) to two distributions, but legend is more clear.
+
+    TODO: Need to align histogram bin size between groups to make the visual
+          representation more interpretable
+
+    **data -- Datasets, name will be used to build labels.
+    bin_rule -- Determines bins width using this function.
+                Will received a joint dataframe of all data passed
+    xrefs - List of values in the bin-range to highlight as vertical lines
+    yrefs - List of values in the count-range to highlight as horizontal lines
+    bins - Number of bins to divide into
+    """
+
+    schema = vega.load_schema("histogram_static_bins_multi.vg.json")
+
+    def hist(label, subset, edges):
+        assignments = np.digitize(subset, edges) - 1
+        counts = np.bincount(assignments)
+        spans = [*(zip(edges, edges[1:]))]
+        desc = [
+            {"bin0": l.item(), "bin1": h.item(), "count": c.item(), "label": label}
+            for ((l, h), c) in zip(spans, counts)
+        ]
+        return desc
+
+    def as_value_list(label, data):
+        try:
+            return data["state_values"].rename(label)
+        except BaseException:
+            return pd.Series(data).rename(label)
+
+    data = {k: as_value_list(k, subset) for k, subset in data.items()}
+
+    joint = pd.concat(data)
+    bins_count = bin_rule(joint)
+    _, edges = np.histogram(joint, bins=bins_count)
+
+    hists = {k: hist(k, subset, edges) for k, subset in data.items()}
+    desc = [item for sublist in hists.values() for item in sublist]
+
+    schema["data"] = vega.replace_named_with(schema["data"], "binned", ["values"], desc)
+
+    schema["data"] = vega.replace_named_with(
+        schema["data"], "xref", ["values"], [{"value": v} for v in xrefs]
+    )
+
+    schema["data"] = vega.replace_named_with(
+        schema["data"], "yref", ["values"], [{"count": v} for v in yrefs]
+    )
+
+    if return_bins:
+        return schema, pd.DataFrame(desc).set_index(["bin0", "bin1"])
+    else:
+        return schema
