+++ src/pyciemss/visuals/schemas/barycenter_triangle.vg.json
