+++ src/pyciemss/visuals/schemas/histogram_static_bins_multi.vg.json
