+++ src/pyciemss/visuals/schemas/multigraph.vg.json
