+++ src/pyciemss/visuals/schemas/spring_graph.vg.json
@@ -1,0 +1,178 @@
+{
+    "$schema": "https://vega.github.io/schema/vega/v5.json",
+    "description": "A node-link diagram with force-directed layout.",
+    "width": 300,
+    "height": 300,
+    "padding": 0,
+  
+    "signals": [
+      { "name": "cx", "update": "width / 2" },
+      { "name": "cy", "update": "height / 2" },
+      { "name": "nodeRadius", "value": 15},
+      { "name": "nodeCharge", "value": -80},
+      { "name": "linkDistance", "value": 80},
+      { "name": "static", "value": true,
+        "bind": {"input": "checkbox"} },
+      {
+        "description": "State variable for active node fix status.",
+        "name": "fix", "value": false,
+        "on": [
+          {
+            "events": "symbol:mouseout[!event.buttons], window:mouseup",
+            "update": "false"
+          },
+          {
+            "events": "symbol:mouseover",
+            "update": "fix || true"
+          },
+          {
+            "events": "[symbol:mousedown, window:mouseup] > window:mousemove!",
+            "update": "xy()",
+            "force": true
+          }
+        ]
+      },
+      {
+        "description": "Graph node most recently interacted with.",
+        "name": "node", "value": null,
+        "on": [
+          {
+            "events": "symbol:mouseover",
+            "update": "fix === true ? item() : node"
+          }
+        ]
+      },
+      {
+        "description": "Flag to restart Force simulation upon data changes.",
+        "name": "restart", "value": false,
+        "on": [
+          {"events": {"signal": "fix"}, "update": "fix && fix.length"}
+        ]
+      }
+    ],
+  
+    "scales": [
+      {
+        "name": "color",
+        "type": "ordinal",
+        "domain": {"data": "node-data", "field": "group"},
+        "range": {"scheme": "category20c"}
+      }
+    ],
+  
+    "marks": [
+      {
+        "name": "nodes",
+        "type": "symbol",
+        "zindex": 1,
+  
+        "from": {"data": "node-data"},
+        "on": [
+          {
+            "trigger": "fix",
+            "modify": "node",
+            "values": "fix === true ? {fx: node.x, fy: node.y} : {fx: fix[0], fy: fix[1]}"
+          },
+          {
+            "trigger": "!fix",
+            "modify": "node", "values": "{fx: null, fy: null}"
+          }
+        ],
+  
+        "encode": {
+          "enter": {
+            "fill": {"scale": "color", "field": "group"},
+            "stroke": {"value": "white"}
+          },
+          "update": {
+            "size": {"signal": "2 * nodeRadius * nodeRadius"},
+            "cursor": {"value": "pointer"}
+          }
+        },
+  
+        "transform": [
+          {
+            "type": "force",
+            "iterations": 300,
+            "restart": {"signal": "restart"},
+            "static": {"signal": "static"},
+            "signal": "force",
+            "forces": [
+              {"force": "center", "x": {"signal": "cx"}, "y": {"signal": "cy"}},
+              {"force": "collide", "radius": {"signal": "nodeRadius"}},
+              {"force": "nbody", "strength": {"signal": "nodeCharge"}},
+              {"force": "link", "links": "link-data", "distance": {"signal": "linkDistance"}}
+            ]
+          }
+        ]
+      },
+      {
+        "type": "path",
+        "from": {"data": "link-data"},
+        "interactive": false,
+        "encode": {
+          "update": {
+            "stroke": {"value": "#ccc"},
+            "strokeWidth": {"value": 0.5}
+          }
+        },
+        "transform": [
+          {
+            "type": "linkpath",
+            "require": {"signal": "force"},
+            "shape": "line",
+            "sourceX": "datum.source.x", "sourceY": "datum.source.y",
+            "targetX": "datum.target.x", "targetY": "datum.target.y"
+          }
+        ]
+      },
+      {
+        "type": "text",
+        "name": "labels",
+        "from": {"data": "nodes"},
+        "transform" : [
+           {"type": "formula", "as": "y", "expr": "datum.y + nodeRadius"}
+        ],
+        "encode": {
+          "enter": {
+            "fill": {"value": "black"},
+            "align": {"value": "center"},
+            "baseline": {"value": "top"},
+            "fontSize": {"value":15}
+          },
+          "update": {
+            "text": {"field": "datum.label"},
+            "x": {"field": "x"},
+            "y": {"field": "y"}
+          }
+        }
+      }      
+    ],
+
+    "data": [
+      {
+        "name": "node-data",
+        "values": [
+          {"id": 0, "label": "Zero"},
+          {"id": 1, "label": "One"},
+          {"id": 2, "label": "Two"},
+          {"id": 3, "label": "Three"},
+          {"id": 4, "label": "Four"},
+          {"id": 5, "label": "Five"},
+          {"id": 6, "label": "Six"}
+        ]
+      },
+      {
+        "name": "link-data",
+        "values": [
+          {"id": 1, "source": 0, "target": 1},
+          {"id": 2, "source": 0, "target": 2},
+          {"id": 3, "source": 0, "target": 3},
+          {"id": 4, "source": 0, "target": 4},
+          {"id": 5, "source": 0, "target": 5},
+          {"id": 6, "source": 0, "target": 6}
+        ]
+      }
+    ]
+  }
+
