+++ src/pyciemss/visuals/schemas/trajectories.vg.json
